requirements.txt
flask
pandas
joblib
werkzeug
scikit-learn
gunicorn
python-dotenv